+++ exporter/opentelemetry-exporter-opencensus/src/opentelemetry/exporter/opencensus/trace_exporter/__init__.py
@@ -46,7 +46,10 @@ class OpenCensusSpanExporter(SpanExporter):
     """
 
     def __init__(
-        self, endpoint=DEFAULT_ENDPOINT, host_name=None, client=None,
+        self,
+        endpoint=DEFAULT_ENDPOINT,
+        host_name=None,
+        client=None,
     ):
         tracer_provider = trace.get_tracer_provider()
         service_name = (